+++ priv/repo/migrations/20170728133150_add_kind_to_user.exs
@@ -1,0 +1,9 @@
+defmodule Icta.Repo.Migrations.AddKindToUser do
+  use Ecto.Migration
+
+  def change do
+    alter table(:users) do
+      add :kind, :string, default: "user", null: false
+    end
+  end
+end